requirements.txt
bidict==0.23.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dominate==2.9.1
Flask==3.0.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.5.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
simple-websocket==1.0.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
visitor==0.1.3
Werkzeug==3.0.2
wsproto==1.2.0
WTForms==3.1.2
